requirements.txt
# Repair Manual Scraper Dependencies
requests>=2.31.0
pyyaml>=6.0
